change IPAIYamlJobConfig location
